requirements.txt
Flask
Jinja2
Flask-SQLAlchemy
Flask-Migrate
PyMySQL
Flask-Login
Flask-WTF
Flask-Bcrypt
email-validator
Flask-Mail
sendgrid>=6.11.0
cffi
python-dateutil
gunicorn
python-dotenv
Flask-JWT-Extended
google-cloud-storage
google-cloud-core
google-api-core
Flask-Limiter
pytz
google-cloud-documentai
Flask-Cors>=4.0.0
firebase-admin==7.1.0
beautifulsoup4==4.13.4
lxml==5.4.0
soupsieve==2.7
flasgger